Added support for NPM package manager
