+++ lib/curated/utils.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require_relative 'utils/github'
+require_relative 'utils/string'
+require_relative 'utils/npm'
+
+module Utils
+end
