+++ lib/curated/utils/string.rb
@@ -6,19 +6,6 @@ require 'json'
 module Utils
   extend self
 
-  def contributors_for_file(repo, path)
-    uri = URI("https://api.github.com/repos/#{repo}/commits?path=#{path}&per_page=20")
-
-    response = Net::HTTP.get(uri)
-    commits = JSON.parse(response)
-
-    commits.map { |commit| commit['author']['login'] }.uniq
-  end
-
-  def github_link_to_login(login)
-    "https://github.com/#{login}"
-  end
-
   def underscore(camel_cased_word)
     return camel_cased_word unless /[A-Z-]|::/.match?(camel_cased_word)
